fix(web10-social): proxy-based lucide-react mock — kills manual icon lists that reddened dev
Failure class: hand-maintained mock icon lists. #322 added Eye/EyeOff to
PostLightbox; six test files mock lucide-react with manual name lists, so
dev went red and blocked PR #331. Shared helpers/lucideMock.tsx fabricates
any icon via Proxy (then/catch/finally guarded so vitest's thenable check
doesn't misfire). 344 tests green, tsc clean.
