+++ CHANGELOG.md
@@ -7,6 +7,8 @@ fix(web10-social): D-ledger-mirror-fix — the public ledger mirror was dead for
 feat(web10-social): D-settings-tab bite A (shell) — /settings route + nav entry + Account section (username/provider read-only, profile link, log out) + Data explainer (terms editor link, export pointer) + About (version, Report a bug reuse, manifesto link). New `components/Settings/SettingsScreen.tsx`, `data/settings.ts` (stub for bite B posting defaults). `.context/settings-tab-proposal.md` with the full Facebook-style tree for operator sign-off. Screenshot harness extended to capture settings at desktop + 375px. 330 tests green, vite build clean, screenshots desktop + 375px.
 1.0.194 || 27.07.2026
 chore: unbrick — false-"pre-existing" attribution loop. A workhorse workspace stalled after its own new test code failed CI typecheck (TS2352 on `as Response` casts in feed.test.ts): it asserted the errors were pre-existing without running tsc on dev, then looped on "let me verify" until the context died. Durable fixes, code-first: (1) `scripts/ci-failures.sh <pr>` — one command that always lands on the failing job's error lines (kills `gh run view --log` archaeology and guess-at-the-cause); (2) `bun run check` in web10-social (tsc --noEmit + vitest run) — one local command matching CI so "did I break typecheck?" is a 10-second answer; (3) markdown: AGENTS.md now forbids unverified "pre-existing" claims (quote origin/dev output or it didn't happen) and CLAUDE.md's `unbrick!` states the bricked workspace is disposable — the fix targets future workspaces, never a rescue of the dead one. Also pushed the actual cast fix to PR #325 (`as unknown as Response`, 337 tests green, tsc clean).
+1.0.195 || 27.07.2026
+fix(web10-social): unbrick, second failure class — hand-maintained lucide-react mock icon lists. #322 added Eye/EyeOff icons to PostLightbox; the six test files that mock lucide-react each keep a manual list of icon names, none had Eye, so dev went red and blocked PR #331 with `No "Eye" export is defined on the "lucide-react" mock`. Structural fix: shared `src/__tests__/helpers/lucideMock.tsx` — a Proxy that fabricates ANY icon on demand (with the `then`/`catch`/`finally` guard so vitest's thenable detection doesn't mistake the mock for a promise). All six test files now do `vi.mock('lucide-react', () => lucideMock)`; adding a new icon can never break a mock again. 344 tests green, tsc clean.
 
 1.0.193 || 27.07.2026
 feat(marketing-ui): D-marketing-search bite a — trending search bar. YouTube-style search bar in the /trending header row (⌘K shortcut, clear button, debounced 300ms). Queries hit PATCH /discover/search (anon-first, no auth needed). Results render in the existing TrendingCard grid with skeleton loading states and a designed empty state ("Nothing matches..." with clear-search button). #tag queries strip the hash and search by tag text; topic chips filter search results. All result clicks deep-link to web10-social. No nav changes, no user results (bite b). 103 marketing-ui tests green, tsc -b + vite build clean.
